% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydelta.R
\name{cases_ext}
\alias{cases_ext}
\title{Extract variables and their names from the formula}
\usage{
cases_ext(formula, mean_dta = NULL, cov_dta = NULL)
}
\arguments{
\item{formula}{A formula object specifying the variables of interest.}

\item{mean_dta}{Vector containing the means of the variables.}

\item{cov_dta}{Covariance matrix of the variables.}
}
\value{
list containing objects with variables and formula
}
\description{
Extract variables and their names from the formula
}
