% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydelta.R
\name{ext_bd_var}
\alias{ext_bd_var}
\title{Extract variables from a formula}
\usage{
ext_bd_var(formula)
}
\arguments{
\item{formula}{A formula object or a character string representing a formula.}
}
\value{
A named character vector of extracted variables.
}
\description{
Extracts variables from a formula string.
}
