% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydelta.R
\name{for_to_exp}
\alias{for_to_exp}
\title{Convert a formula to an expression}
\usage{
for_to_exp(formula)
}
\arguments{
\item{formula}{A formula object or a character string representing a formula.}
}
\value{
The evaluated expression.
}
\description{
Converts a formula to an expression for further evaluation.
}
