% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydelta.R
\name{tidydelta_m}
\alias{tidydelta_m}
\title{Estimate means and standard errors}
\usage{
tidydelta_m(
  formula,
  normality_eval = T,
  formula_vars = mean,
  mean_dta = NULL,
  cov_dta = NULL,
  n = NULL,
  conf_lev = 0.95
)
}
\arguments{
\item{formula}{A formula object specifying the variables of interest.}

\item{normality_eval}{Logical value to run normality test in case of being
possible.}

\item{formula_vars}{The function(s) to apply to the variables in the formula.}

\item{mean_dta}{Data frame containing the means of the variables.}

\item{cov_dta}{Covariance matrix of the variables.}

\item{n}{Sample size evaluate (in case that we can evaluate the confidence
intervals with different hypnotic sample sizes).}

\item{conf_lev}{Confidence level for confidence intervals.}
}
\value{
A tibble with columns for means, standard errors, and optionally, confidence intervals.
}
\description{
Estimates means and standard errors for a given formula using Monte Carlo simulation.
}
