% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydelta.R
\name{where_env}
\alias{where_env}
\title{Recursive search of environment}
\usage{
where_env(name, env = rlang::caller_env())
}
\arguments{
\item{name}{Object searched}

\item{env}{Initial environment to search}
}
\value{
A named character vector of extracted variables.
}
\description{
Recursive search of environment containing object.
}
